ml/pipelines/training/requirements.txt
mlflow==2.12.1
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.2
xgboost==2.0.3
torch==2.2.1
pydantic==2.6.3
pyyaml==6.0.1
